Begin Associations param handler refactors
