+++ app/classes/query/location_with_observations.rb
@@ -42,7 +42,8 @@ module Query
 
     def initialize_association_parameters
       initialize_projects_parameter
-      add_for_project_condition
+      add_for_project_condition(:project_observations,
+                                [:observations, :project_observations])
       initialize_species_lists_parameter
       add_in_species_list_condition
       initialize_herbaria_parameter
